+++ src/botlib/be_ai_move.cpp
@@ -606,7 +606,7 @@ float DistanceFromLineSquared(vec3_t p, vec3_t lp1, vec3_t lp2)
 			(proj[j] < lp1[j] && proj[j] < lp2[j]))
 			break;
 	if (j < 3) {
-		if (fabs(proj[j] - lp1[j]) < fabs(proj[j] - lp2[j]))
+		if (fabsf(proj[j] - lp1[j]) < fabsf(proj[j] - lp2[j]))
 			VectorSubtract(p, lp1, dir);
 		else
 			VectorSubtract(p, lp2, dir);
@@ -1269,7 +1269,7 @@ void BotCheckBlocked(bot_movestate_t *ms, vec3_t dir, int checkbottom, bot_mover
 	//test for entities obstructing the bot's path
 	AAS_PresenceTypeBoundingBox(ms->presencetype, mins, maxs);
 	//
-	if (fabs(DotProduct(dir, up)) < 0.7)
+	if (fabsf(DotProduct(dir, up)) < 0.7f)
 	{
 		mins[2] += sv_maxstep->value; //if the bot can step on
 		maxs[2] -= 10; //a little lower to avoid low ceiling
@@ -2224,7 +2224,7 @@ bot_moveresult_t BotFinishTravel_Elevator(bot_movestate_t *ms, aas_reachability_
 	//
 	VectorSubtract(reach->end, ms->origin, topdir);
 	//
-	if (fabs(bottomdir[2]) < fabs(topdir[2]))
+	if (fabsf(bottomdir[2]) < fabsf(topdir[2]))
 	{
 		VectorNormalize(bottomdir);
 		EA_Move(ms->client, bottomdir, 300);
@@ -2674,8 +2674,8 @@ bot_moveresult_t BotTravel_Grapple(bot_movestate_t *ms, aas_reachability_t *reac
 		result.flags |= MOVERESULT_MOVEMENTVIEW;
 		//
 		if (dist < 5 &&
-			fabs(AngleDiff(result.ideal_viewangles[0], ms->viewangles[0])) < 2 &&
-			fabs(AngleDiff(result.ideal_viewangles[1], ms->viewangles[1])) < 2)
+			fabsf(AngleDiff(result.ideal_viewangles[0], ms->viewangles[0])) < 2 &&
+			fabsf(AngleDiff(result.ideal_viewangles[1], ms->viewangles[1])) < 2)
 		{
 #ifdef DEBUG_GRAPPLE
 			botimport.Print(PRT_MESSAGE, "BotTravel_Grapple: activating grapple\n");
@@ -2743,8 +2743,8 @@ bot_moveresult_t BotTravel_RocketJump(bot_movestate_t *ms, aas_reachability_t *r
 	result.ideal_viewangles[PITCH] = 90;
 	//
 	if (dist < 5 &&
-			fabs(AngleDiff(result.ideal_viewangles[0], ms->viewangles[0])) < 5 &&
-			fabs(AngleDiff(result.ideal_viewangles[1], ms->viewangles[1])) < 5)
+			fabsf(AngleDiff(result.ideal_viewangles[0], ms->viewangles[0])) < 5 &&
+			fabsf(AngleDiff(result.ideal_viewangles[1], ms->viewangles[1])) < 5)
 	{
 		//botimport.Print(PRT_MESSAGE, "between jump start and run start point\n");
 		hordir[0] = reach->end[0] - ms->origin[0];
@@ -2804,8 +2804,8 @@ bot_moveresult_t BotTravel_BFGJump(bot_movestate_t *ms, aas_reachability_t *reac
 	dist = VectorNormalize(hordir);
 	//
 	if (dist < 5 &&
-			fabs(AngleDiff(result.ideal_viewangles[0], ms->viewangles[0])) < 5 &&
-			fabs(AngleDiff(result.ideal_viewangles[1], ms->viewangles[1])) < 5)
+			fabsf(AngleDiff(result.ideal_viewangles[0], ms->viewangles[0])) < 5 &&
+			fabsf(AngleDiff(result.ideal_viewangles[1], ms->viewangles[1])) < 5)
 	{
 		//botimport.Print(PRT_MESSAGE, "between jump start and run start point\n");
 		hordir[0] = reach->end[0] - ms->origin[0];
